[printer] expand pdf_handle_char c var type to avoid short circuit compare.

+++ lib/printer-emulator/atari_822.cpp
@@ -18,7 +18,7 @@ void atari822::post_new_file()
     pdf_header();
 }
 
-void atari822::pdf_handle_char(uint8_t c, uint8_t aux1, uint8_t aux2)
+void atari822::pdf_handle_char(uint16_t c, uint8_t aux1, uint8_t aux2)
 {
     // use PDF inline image to display line of graphics
     /*
